rag_loan_chatbot_final/requirements.txt
pandas
numpy
sentence-transformers
faiss-cpu
transformers
torch
streamlit